# server/docs/swagger.yaml
---
swagger: '2.0'
info:
  version: 1.0.0
  title: PQVP
  description: |
    #### API for PQVP Emergency Response Prototype
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
host: demo.pqvp.truss.works
basePath: /api
paths:
  /login:
    post:
      description: Logs user into the application
      produces:
        - application/json
      responses:
        200:
          description: Login Successful
        400:
          description: Invalid Request
        401:
          description: Authorization Failed
        500:
          description: Internal Server Failure
      parameters:
        - name: email
          in: query
          description: users email address to login
          type: string
          required: true
        - name: password
          in: query
          description: users password to login
          type: string
          required: true
  /signup:
    post:
      description: Creates a new user account
      produces:
        - application/json
      responses:
        201:
          description: Login Successful
        400:
          description: Invalid Request
        500:
          description: Internal Server Failure
      parameters:
        - name: email
          in: query
          description: email address of user to access account
          type: string
          required: true
        - name: password
          in: query
          description: password to access account
          type: string
          required: true
        - name: admin
          in: query
          description: boolean flag to determine if a user should be an admin (for prototype purposes only, this would be removed for production)
          type: boolean
          required: false
  /profile:
    get:
      description: Returns a logged in user profile
      produces:
        - application/json
      responses:
        200:
          description: Profile returned
        400:
          description: Invalid Request
        400:
          description: Internal Server Failure
    post:
      description: Updates a logged in user profile
      produces:
        - application/json
      responses:
        200:
          description: Profile updated
        400:
          description: Invalid Request
        500:
          description: Internal Server Failure
      parameters:
        - name: phone
          in: query
          description: phone number of the user
          type: string
          required: true
        - name: addresses
          in: query
          description: Array of address, latitude, longitude
          type: array
          required: true